# src/main/java/com/nuevo/taskmanager/dto/LoginResponse.java
package com.nuevo.taskmanager.dto;

public class LoginResponse {
    private String token;

    public LoginResponse() {
    }

    public LoginResponse(String token) {
        this.token = token;
    }

    public String getToken() {
        return token;
    }

    public void setToken(String token) {
        this.token = token;
    }
}